Eliminated unnecessary cast

+++ EPPlus/CellStore.cs
@@ -650,7 +650,7 @@ using OfficeOpenXml;
                         pageItem = _columnIndex[col]._pages[pos];
                     }
 
-                    int ix = (int)(Row - ((pageItem.Index << pageBits) + pageItem.Offset));
+                    int ix = Row - ((pageItem.Index << pageBits) + pageItem.Offset);
                     _searchItem.Index = ix;
                     var cellPos = Array.BinarySearch(pageItem.Rows, 0, pageItem.RowCount, _searchItem);
                     if (cellPos < 0)
@@ -668,7 +668,7 @@ using OfficeOpenXml;
                     col = ~col;
                     AddColumn(col, Column);
                     AddPage(_columnIndex[col], 0, page);
-                    int ix = (int)(Row - (page << pageBits));
+                    int ix = Row - (page << pageBits);
                     AddCell(_columnIndex[col], 0, 0, ix, Value);
                 }
             }
@@ -739,7 +739,7 @@ using OfficeOpenXml;
                                     pageItem = _columnIndex[col]._pages[pos];
                                 }
 
-                                int ix = (int)(rowIx - ((pageItem.Index << pageBits) + pageItem.Offset));
+                                int ix = rowIx - ((pageItem.Index << pageBits) + pageItem.Offset);
                                 _searchItem.Index = ix;
                                 var cellPos = Array.BinarySearch(pageItem.Rows, 0, pageItem.RowCount, _searchItem);
                                 if (cellPos < 0)
@@ -758,7 +758,7 @@ using OfficeOpenXml;
                                 col = ~col;
                                 AddColumn(col, colIx);
                                 AddPage(_columnIndex[col], 0, page);
-                                int ix = (int)(rowIx - (page << pageBits));
+                                int ix = rowIx - (page << pageBits);
                                 AddCell(_columnIndex[col], 0, 0, ix, default(T));
                                 Updater(_values, _columnIndex[col]._pages[0].Rows[0].IndexPointer, rowIx, colIx, Value);
                             }
@@ -809,7 +809,7 @@ using OfficeOpenXml;
                         pageItem = _columnIndex[col]._pages[pos];
                     }
 
-                    int ix = (int)(Row - ((pageItem.Index << pageBits) + pageItem.Offset));
+                    int ix = Row - ((pageItem.Index << pageBits) + pageItem.Offset);
                     _searchItem.Index = ix;
                     var cellPos = Array.BinarySearch(pageItem.Rows, 0, pageItem.RowCount, _searchItem);
                     if (cellPos < 0)
@@ -828,7 +828,7 @@ using OfficeOpenXml;
                     col = ~col;
                     AddColumn(col, Column);
                     AddPage(_columnIndex[col], 0, page);
-                    int ix = (int)(Row - (page << pageBits));
+                    int ix = Row - (page << pageBits);
                     AddCell(_columnIndex[col], 0, 0, ix, default(T));
                     Updater(_values, _columnIndex[col]._pages[0].Rows[0].IndexPointer, Value);
                 }
